Generate corosync.conf.5{,.html} and corosync.aug from single file
This file is conf/corosync.rnc that models set of valid (in relatively
weak sense) corosync configurations as captured in corosync.conf and
uses RELAX NG Compact Syntax notation for that purpose.

This was chosen as the most viable option (e.g., when compared
to standard XML encoding of RELAX NG, see below for more reasons)
for modifications and comprehension performed *by humans* and
its brevity is also importand from versioning system perspective.
For automated processing, it can be (and actually is being done
as one of the step towards a final man page/Augeas lens) easily
converted into computer-friendly XML notation, so the advantage
of both approaches can be utilized.  Big advantage of RELAX NG
is the simple extendability using various annotations and/or
tags from custom namespace.

As a side-effect, this schema can be directly used to validate
XML form of corosync.conf (see COROSYNC.XML(5)) if one prefers
it (and beside that, it is to be used in cman->pcmk cluster
conversion utility, which was a prime reason for this changset
originally).

Apparently, having single authoritative source/model of possible
configurations as a source for producing corosync.conf documentation
and Augeas lens for this file alleviates maintainer's effort and
indirectly leads to more up-to-date/up-to-reality reflection of
possible configuration items.  Systematic approach is usually
pretty reasonable...

The core of this changeset, corosync.rnc, is mostly a result
of copy-paste from original man page, analyzing coroparse.c
and few others for the expected types/sets of values for
particular options and some other annotation work (see
integration task 4/ below).

Technical details
-----------------

- generating corosync.conf documentation:

                  rnctree.py
-> [corosync.rnc] ---------> [corosync.orig.rng] --------+
                                                         |
                                                     xsltproc
                                            simplification-wrapper.xsl
                          xsltproc                       |
[corosync.conf.asciidoc] <-------- [corosync.rng] <------+
       +            rng2asciidocman.xsl
       |                               xsltproc
    asciidoc                        +-----------> [corosync.conf.5]
       |                           / manpage.xsl
       +--> [corosync.conf.xml] --K
                                   \  xsltproc
                                    +-----------> [corosync.conf.5.html]
                                      xhtml.xsl

- generating corosync.aug Augeas lens:

                  rnctree.py
-> [corosync.rnc] ---------> [corosync.orig.rng] --------+
                                                         |
                                                     xsltproc
                                                   rng2lens.xsl
                                                         |
                                   [corosync.aug] <------+

- apparently, new dependencies in build-time are following
  (checks also added into configure.ac):
  - xsltproc (or another XSLT 1.0 compliant processor)
  - DocBook 5+ stylesheets
  - AsciiDoc, and therefore
    - Python (version [2.6, 3.0] should be fine)
  - optionally trang, but custom rnc2rng (see below) logic is embedded
    (this requires Python as well)

- it might not be difficult to generate C parser code based on the
  schema described in corosync.rnc in a similar way in the future,
  however this is the next-level challenge

Licensing concerns and bundling files of other projects
-------------------------------------------------------

Not all the added files are authored by me, however they are either
licensed also under revised BSD (lex.py, rnc_tokenize.py, rnctree.py)
or I got approval from the original author as the original licensing
was unclear (simplification.xsl).  See LICENSE file modifications.

The paragraph above enumerated the files that were bundled due to
a lack of respective packages in RHEL or elsewhere and partly because
they are used solely for the purpose or producing a build/release.

The respective projects are:
- rnc2rng [1] (rnc_tokenize.py, rnctree.py)
  - this is a light-weight alternative to trang
  - originally public domain, I maintain the extended version licensed
    under revised BSD (with an awareness of the original author)
    and fined-tuned particularly for this purpose
- PLY [2] (lex.py)
  - lexical analyzer used in rnc2rng, which bundles it as well
    "since beginning"
- "relax-ng utitilites" [3] (simplification.xsl)

Integration tasks
-----------------

The changeset needs some more love in several aspects:

1/ which files are to be part of installation/distribution/tarball
   (see ascii-art schemes under Technical details)
   - corosync.aug, corosync.conf.5{,.html}?
   - corosync.rng?

2/ autotools integration
   I am not an expert here;  the dependencies based on modification
   time seem to work only occasinally, "make clear" misses some
   by-products, "make dist" and the like may forget some files
   (see also 1/)

3/ portability
   - will AsciiDoc always install XSLT wrappers in
     /etc/asciidoc/docbook-xsl?
   - and many others, also the variability in locations is connected
     to 2/

4/ reflect current state in corosync.rnc
   - I tried hard to make sure the schema corresponds to the reality
     (using original man page and Augeas lens, and the code itself)
     but it should be reviewed carefully by the corosync expert

[1] http://fedorapeople.org/cgit/jpokorny/public_git/rnc2rng.git/
    (real main project homepage is expected soon)
[2] http://www.dabeaz.com/ply/
[3] http://downloads.xmlschemata.org/relax-ng/utilities/

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>