Add timestamps to the Corosync black-box output
This patch adds timestamps to the flight-data file so that its output
can be correlated with syslog to aid in debugging. It also adds a header to
the file so that corosync-fplay can tell whether it was generated by the new
code or the old code.

So corosync-fplay in this code can read new and old format fdata files. The old corosync-fplay will crash when trying to read new format files.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>
Reviewed-By: Jan Friesse <jfriesse@redhat.com>