logsys: Handle full /dev/shm correctly
If /dev/shm was full, corosync was segfaulting instead of display
error and exit with proper exit code (code > 127 is reserved for
signals).

Also when logsys is not initialized, error message was displayed
only on stderr. This is not ok if corosync is running as service
and solution is to send error also to syslog.

Signed-off-by: Jan Friesse <jfriesse@redhat.com>
Reviewed-by: Christine Caulfield <ccaulfie@redhat.com>