Add clang-format configuration file
This .clang-format file is written for clang-format version 3.7.1

I've attempted to set the options for clang-format so that the
difference between the current code, and the result of the clang format
call is as small as possible.

Unfortunately, clang-format doesn't yet have the ability to handle every
single possible formatting option, so it's not perfect yet.

Signed-off-by: Michael Jones <jonesmz@jonesmz.com>
Reviewed-by: Jan Friesse <jfriesse@redhat.com>