Spec: make corosynclib-devel usable without pulling corosync
It's expected that libraries are still working properly (e.g. indicating
errors to their users) when there's no corosync process around in that
moment, and from this perspective it doesn't matter whether it is
installed at all for some purposes, especially having linkage with them
in mind.

Note that the inverse dependency, main corosync package on corosynclib,
is already there (not strictly needed, likely just to enforce package
version match -- otherwise RPM's dependency generator will handle this
on its own using SONAMEs -- hence the comments to that effect are also
added), so breaking this symmetry:
- is supposed to be harmless modulo cases that should be fixed to
  express explicit dependency on corosync's runtime anyway
  (but only for runtime, i.e., Requires as opposed to BuildRequires)
- will effectively enable more lightweight get-build-deps-and-build
  process for programs linking with corosynclibs (e.g. pacemaker),
  as corosync package won't need to be installed needlessly

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>