totemrrp: Fix situation when all rings are faulty
Previously when all rings were marked as failed no message was sent via
any interface. This is problem because totemsrp relies on messages
delivered via localhost multicast loop socket so it never moved to
single node mode and it got stuck.

Solution is to send message via one of the interfaces (first one seems
to be obvious choice) if it was not sent via any interface.

Strictly speaking it should be enough to change just *_mcast_flush_send
functions, but changing others is just for sure.

Signed-off-by: Jan Friesse <jfriesse@redhat.com>