add mechanism to trigger interface up and down for udp transportation
For https://github.com/corosync/corosync/pull/32, there is some issue
with corosync, which is that on the node we execute "ifconfig ethX down"
we can still see that the other node is still online. This patch could
trigger interface status change event, which make that you can see the
other node is offline when you execute "ifconfig eth0 down" on this
node.