votequorum: make auto-tie-breaker consistent on cluster membership changes
When the cluster changes from even sized to odd sized corosync disables auto-tie-breaker if wait_for_all is not enabled.
However when changing from odd sized to even sized it doesn't reenable it, causing auto_tie_breaker to be inconsistent across the cluster:
the newly added node and any nodes that restart corosync will have it, but all the previously running nodes won't.

Signed-off-by: Edwin Torok <edvin.torok@citrix.com>