corosync.conf: publicize nodelist.node.name configuration option
It was discovered that pacemaker has been occassionaly relying on
those items configured in corosync.conf (and documenting so), while
backpropagation got stuck somewhere.  As the option is deemed generally
beneficial by Honza[1], rectify this gap now and make it standard,
public part of the configuration space, possibly also for other
client SW to use now.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1517834#c6

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>