SECURITY: be consistent on the hash algorithm used
The SECURITY doc mentioned both SHA1 and SHA256. Although this is configurable, the documentation should be consistent on what the default actually is.
By looking at the `encrypt_and_sign` function it used `instance->crypto_hash_type` which is set in `totem_get_crypto`, defaulting to `sha1` if not overriden.