Send msgs to ourselves via localhost when in loopback mode
Whenever the interface went down, or the interface lost the IP, the
loopback mode was not working - it was still sending messages to its own
IP, and listening on loopback, so it would never receive its own
messages.

Corosync in UDPU mode would realise it has lost contact to everyone else
when an interface is taken down or the IP address lost, but would fail
to form a new membership because it was not even able to talk to itself.
This is dangerous because there could be another interface or path to
the storage still up and might corrupt GFS2 data. When the interface
got brought up corosync would then dump core with an assertion failure.
This is a well known problem, see
https://bugzilla.redhat.com/show_bug.cgi?id=1397923 and
https://bugzilla.redhat.com/show_bug.cgi?id=989934.

This patch ensures that we send messages to localhost in this case,
instead of our IP, both for unicast and multicast messages (both
these changes are necessary for it to work).

After this patch corosync will be able to talk to itself (which is what
the intent of the loopback mode seems to have been), form a new
membership just with itself, see that it doesn't have quorum and
eventually SBD will fence the node as it should.

All credit goes to @edwintorok for finding the cause of this issue and
the solution.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>